added avg Grade function
