+++ client/src/components/App/NavMenu/Assignments.jsx
@@ -28,7 +28,8 @@ const Assignments = () => {
     { title: "Task K", submitted: true, comments: "See me after class" },
     { title: "Task L", submitted: false, comments: "This was a good attempt" },
   ];
-
+  
+  
   
   return (
     <>
