+++ client/src/components/App/NavMenu/Grades.jsx
@@ -1,27 +1,8 @@
 import React, { useContext } from "react";
 import LandingPageContext from "../LandingPage/LandingPageContext";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faCheck } from "@fortawesome/free-solid-svg-icons";
 const Grades = () => {
-  const { isDarkMode } = useContext(LandingPageContext);
-  const tableData = [
-    { title: "Project A", grade: "93%" },
-    { title: "Test Alpha", grade: "67%" },
-    { title: "Essay I", grade: "88%" },
-    { title: "Quiz X", grade: "99%" },
-    { title: "Homework 5", grade: "91%" },
-    { title: "Exam II", grade: "77%" },
-    { title: "Lab 3", grade: "70%" },
-    { title: "Report Z", grade: "84%" },
-    { title: "Task 4", grade: "94%" },
-    { title: "Module Test", grade: "92%" },
-    { title: "Presentation 1", grade: "100%" },
-    { title: "Fieldwork", grade: "79%" },
-    { title: "Study II", grade: "86%" },
-    { title: "Workshop 3", grade: "76%" },
-    { title: "Activity B", grade: "89%" },
-];
-
+  const { isDarkMode, tableData } = useContext(LandingPageContext);
+  
   return (
     <>
       <div
